requirements.txt
flask==3.0.3
line-bot-sdk==3.11.0
anthropic==0.40.0
gspread==6.1.4
google-auth==2.35.0
python-dotenv==1.0.1
schedule==1.2.2
gunicorn==23.0.0
